requirements.txt
flask
ultralytics
jinja2
opencv-python
numpy
pillow
torch
torchvision
